Refactor and emit fixes
