+++ examples/lexicons/codelingo/ast/typescript/src/main.ts
@@ -1,14 +1,14 @@
 import { AstNode } from "./model";
 import { parseProject } from "./Parser";
 
-const printNode = (node: AstNode) => {}; //console.log(JSON.stringify(node));
+const emitNode = (node: AstNode) => console.log(JSON.stringify(node));
 
 function main(args: string[]) {
     const [_, __, method, ...rest] = args;
     switch (method) {
         case "parse-project":
             const [trunkKey, baseDir, ...filepaths] = rest;
-            parseProject(trunkKey, baseDir, filepaths, printNode);
+            parseProject(trunkKey, baseDir, filepaths, emitNode);
             break;
     }
 }
